+++ localDevAPI/main.py
@@ -1,86 +1,19 @@
 from typing import Union, Annotated
 
 import os
-import os.path as osp
 import sys
-from datetime import datetime, timedelta, timezone
+from datetime import timedelta
 
 from fastapi import FastAPI, Depends, HTTPException, status
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
-from passlib.context import CryptContext
-import jwt
-from jwt.exceptions import InvalidTokenError
+from fastapi.security import OAuth2PasswordRequestForm
 
 file_dir = os.path.dirname(__file__)
 root_dir = os.path.dirname(file_dir)
 sys.path.insert(0, root_dir)
-from security.classstructures import Token, TokenData, Item, User, UserInDB
+from security.classstructures import Token, Item, User, UserInDB
+from security.authentication import *
+from security import ACCESS_TOKEN_EXPIRE_MINUTES
 
-#Example of accessing the secrets within the container
-secret_path = 'run/secrets/oauth_token'
-
-if osp.isfile(secret_path):
-    with open('/run/secrets/oauth_token', 'r') as f:
-            oauth_token = f.read().strip()
-else:
-    # to get a string like this run:
-    # openssl rand -hex 32
-    SECRET_KEY = os.getenv('SECRET_KEY')
-    ALGORITHM = "HS256"
-    ACCESS_TOKEN_EXPIRE_MINUTES = 30
-
-fake_users_db = {
-    "johndoe": {
-        "username": "johndoe",
-        "full_name": "John Doe",
-        "email": "johndoe@example.com",
-        "hashed_password": "$2b$12$EixZaYVK1fsbw1ZfbX3OXePaWxn96p36WQoeG6Lruj3vjPGga31lW",
-        "disabled": False,
-    },
-    "alice": {
-        "username": "alice",
-        "full_name": "Alice Wonderson",
-        "email": "alice@example.com",
-        "hashed_password": "fakehashedsecret2",
-        "disabled": True,
-    },
-}
-
-def fake_hash_password(password: str):
-    return "fakehashed" + password
-
-def get_user(db, username:str):
-    if username in db:
-        user_dict = db.get(username)
-        return UserInDB(**user_dict)
-
-def verify_password(plain_password, hashed_password):
-    return pwd_context.verify(plain_password, hashed_password)
-
-def get_password_hash(password):
-    return pwd_context.hash(password)
-
-def authenticate_user(fake_db, username: str, password: str):
-    user = get_user(fake_db, username)
-    if not user:
-        return False
-    if not verify_password(password, user.hashed_password):
-        return False
-    return user
-
-def create_access_token(data: dict, expires_delta: timedelta | None = None):
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.now(timezone.utc) + expires_delta
-    else:
-        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
-
-# Oauth
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl='token')
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
 app = FastAPI(
     title="Web Application with Authentication",
@@ -108,36 +41,6 @@ def read_item(item_id: int, q: Union[str, None] = None):
 @app.put("/items/{item_id}")
 def update_item(item_id: int, item: Item):
     return {"item_name": item.name, "item_id": item_id}
-
-def fake_decode_token(token):
-    """This does not provide any security yet"""
-    return get_user(fake_users_db, token)
-
-async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)]):
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate crendentials",
-        headers={"WWW-Authenticate": "Bearer"}
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        token_data = TokenData(username=username)
-    except InvalidTokenError:
-        raise credentials_exception
-    user = get_user(fake_users_db, username=token_data.username)
-    if user is None:
-        raise credentials_exception
-    return user 
-
-async def get_current_active_user(
-    current_user: Annotated[User, Depends(get_current_user)],
-):
-    if current_user.disabled:
-        raise HTTPException(status_code=400, detail="Inactive user")
-    return current_user
 
 @app.post("/token")
 async def login_for_access_token(
